doc2jpg/requirements.txt
PyMuPDF==1.21.1
Pillow==9.4.0
python-docx2pdf==0.1.8 